made own file for classes
