+++ lib/joshuawidgets/listviewjjc.dart
@@ -1,0 +1,34 @@
+import 'package:flutter/material.dart';
+
+class MyListView extends StatelessWidget {
+  const MyListView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    const title = "Basic List";
+    return MaterialApp(
+      title: title,
+      home: Scaffold(
+        appBar: AppBar(
+          title: const Text(title),
+        ),
+        body: ListView(
+          children: const <Widget>[
+            ListTile(
+              leading: Icon(Icons.map),
+              title: Text('Map'),
+            ),
+            ListTile(
+              leading: Icon(Icons.album),
+              title: Text('Album'),
+            ),
+            ListTile(
+              leading: Icon(Icons.phone),
+              title: Text('Phone'),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
